feat(queue-health): F2b DLQ poll emit + F3/F4/F5 poll-loop wiring (Refs #133)
Per-poll health feed in the processing loop (update_health_probes, called once
per poll after queue-depth metrics):
- A3 DLQ delta-rate: count_dlq() each poll, emit the signed delta via the
  switchboard; prev_dlq seeds from the live count on poll 1 (DOM-09), not 0.
- F5 drain: sample pending bytes into the drain snapshot; emit throughput 0 on
  an idle (no-pending) poll so the lane decays instead of holding a stale rate.
- B4: push the poll-local outcome ring (items_processed/dlq/attempts), compute
  the all-failing predicate, store it in the lock-free atomic.
- A1/A2/A3: evaluate, debounce via observe(), and cache the debounced results
  for the on-RPC verdict.

DLQ control sample retyped u64 -> f64 (sample.rs/scalar_sample/store_dlq_depth):
a draining DLQ is a negative delta-rate, and the A3 probe needs the sign to tell
draining (Green) from growing (Red). reload_baselines drops the DLQ arm
(persist:false, a rate not a level baseline — DATA-10).

EwmaState now carries the QueueHealthConfig (config()), so the poll loop and
verdict read probe thresholds without a separate handle; min_item_bytes() reads
it. QueueManager::count_dlq() + read_disk_space() (sysinfo B2 seam) added.

LoopState gains prev_dlq / dlq_samples_seen / outcome_ring. 116 green.

+++ src/rust/daemon/core/src/switchboard/sample.rs
@@ -46,8 +46,11 @@ pub enum MetricSample {
     /// Per-byte processing cost (ms/KB) — a ratio, carried as `f64` so the EWMA
     /// lane keeps sub-integer precision.
     QueueMsPerKb(f64),
-    /// Dead-letter-queue depth (count).
-    QueueDlqDepth(u64),
+    /// Dead-letter-queue per-poll **delta-rate** (counts/poll), carried as `f64`
+    /// so a *draining* DLQ (negative delta) is preserved — the A3 probe needs the
+    /// sign to distinguish draining (Green) from growing (Red). The lane smooths
+    /// this rate; the absolute count is read separately for the emptiness test.
+    QueueDlqDepth(f64),
     QueueThroughput(f64),
     EmbedderBatch {
         rec: EmbedderBatchRec,
@@ -86,7 +89,10 @@ mod tests {
         };
         assert_eq!(s.id(), MetricId::EmbedderLatency);
         assert_eq!(MetricSample::QueueMsPerKb(7.0).id(), MetricId::QueueMsPerKb);
-        assert_eq!(MetricSample::QueueDlqDepth(7).id(), MetricId::QueueDlqDepth);
+        assert_eq!(
+            MetricSample::QueueDlqDepth(7.0).id(),
+            MetricId::QueueDlqDepth
+        );
         assert_eq!(
             MetricSample::QueueThroughput(1.0).id(),
             MetricId::QueueThroughput
@@ -95,7 +101,7 @@ mod tests {
 
     #[test]
     fn test_sample_is_copy() {
-        let s = MetricSample::QueueDlqDepth(42);
+        let s = MetricSample::QueueDlqDepth(42.0);
         let s2 = s; // Copy — `s` still usable below.
         assert_eq!(s.id(), s2.id());
     }
